Moved 'unzipFile' test to functional
* The test does filesystem changes so it should be tested in e2e
